# docs/API/MessageController-OpenAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Message Center API",
    "description": "消息中心管理系统的完整API接口文档",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "https://api.example.com",
      "description": "生产环境"
    },
    {
      "url": "https://api-dev.example.com",
      "description": "开发环境"
    }
  ],
...
  "paths": {
...
    "/api/messages/{id}": {
      "get": {
        "tags": ["Messages"],
        "summary": "根据ID获取消息",
        "description": "获取指定ID的消息详情",
        "operationId": "getMessage",
        "parameters": [
          {
            "$ref": "#/components/parameters/MessageId"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Messages"],
        "summary": "删除消息",
        "description": "删除指定ID的消息",
        "operationId": "deleteMessage",
        "parameters": [
          {
            "$ref": "#/components/parameters/MessageId"
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer Token认证"
      }
    },
    "parameters": {
      "MessageId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "消息ID"
      },
...
    },
    "schemas": {
...
      "MessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "messageType": {
            "$ref": "#/components/schemas/MessageType"
          },
          "channel": {
            "$ref": "#/components/schemas/MessageChannel"
          },
          "status": {
            "$ref": "#/components/schemas/MessageStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/MessagePriority"
          },
          "senderId": {
            "type": "string",
            "format": "uuid"
          },
          "senderName": {
            "type": "string"
          },
          "receiverId": {
            "type": "string"
          },
          "receiverName": {
            "type": "string"
          },
          "receiverEmail": {
            "type": "string",
            "format": "email"
          },
          "receiverPhone": {
            "type": "string"
          },
          "templateId": {
            "type": "string",
            "format": "uuid"
          },
          "businessType": {
            "type": "string"
          },
          "businessId": {
            "type": "string"
          },
          "scheduledSendTime": {
            "type": "string",
            "format": "date-time"
          },
          "actualSendTime": {
            "type": "string",
            "format": "date-time"
          },
          "deliveredTime": {
            "type": "string",
            "format": "date-time"
          },
          "readTime": {
            "type": "string",
            "format": "date-time"
          },
          "expirationTime": {
            "type": "string",
            "format": "date-time"
          },
          "retryCount": {
            "type": "integer"
          },
          "maxRetryCount": {
            "type": "integer"
          },
          "failureReason": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "tags": {
            "type": "string"
          },
          "linkUrl": {
            "type": "string",
            "format": "uri"
          },
          "attachmentIds": {
            "type": "string"
          },
          "isRead": {
            "type": "boolean"
          },
          "creationTime": {
            "type": "string",
            "format": "date-time"
          },
          "lastModificationTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MessageType": {
        "type": "integer",
        "enum": [1, 2, 3, 4, 5, 6, 7, 8],
        "description": "消息类型：1-通知, 2-工作流, 3-警报, 4-事务, 5-营销, 6-社交, 7-系统, 8-实时"
      },
      "MessageChannel": {
        "type": "integer",
        "enum": [1, 2, 3, 4, 5, 6, 7, 8],
        "description": "消息渠道：1-站内信, 2-邮件, 3-短信, 4-推送, 5-微信, 6-钉钉, 7-WebSocket, 8-外部"
      },
      "MessageStatus": {
        "type": "integer",
        "enum": [1, 2, 3, 4, 5, 6, 7, 8],
        "description": "消息状态：1-待发送, 2-发送中, 3-已发送, 4-已送达, 5-已读, 6-失败, 7-已取消, 8-已过期"
      },
      "MessagePriority": {
        "type": "integer",
        "enum": [1, 2, 3, 4],
        "description": "消息优先级：1-低, 2-普通, 3-高, 4-紧急"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "responses": {
...
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  }
}